add service layer and exception handling

+++ spring/highschool/src/main/java/com/highschool/repository/SubjectRespository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.highschool.model.Subject;
 
 public interface SubjectRespository extends JpaRepository<Subject, Long>{
-    List<Subject> getByTeacherId(int teacherId);
+    List<Subject> getByTeacherId(Long teacherId);
 }
